Refactored project structure for high efficiency and organization (Issue #55)
