+++ .idea/libraries/Dart_SDK.xml
@@ -1,0 +1,30 @@
+<component name="libraryTable">
+  <library name="Dart SDK">
+    <CLASSES>
+      <root url="file://$USER_HOME$/dart-sdk/lib/async" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/cli" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/collection" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/concurrent" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/convert" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/core" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/developer" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/ffi" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/html" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/indexed_db" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/io" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/isolate" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/js" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/js_interop" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/js_interop_unsafe" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/js_util" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/math" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/mirrors" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/svg" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/typed_data" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/web_audio" />
+      <root url="file://$USER_HOME$/dart-sdk/lib/web_gl" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
