+++ .idea/libraries/Flutter_Plugins.xml
@@ -1,0 +1,7 @@
+<component name="libraryTable">
+  <library name="Flutter Plugins" type="FlutterPluginsLibraryType">
+    <CLASSES />
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
