+++ app/lib/features/dashboard/models/activity.dart
