+++ app/lib/features/dashboard/models/course_progress.dart
